Initial commit. Rook, Bishop, Pawn
A lot of the leg work has been done at the time of this commit. Most of the boiler plate and major logic issues are accounted for. Now we just need to start filling in the actual movement templates and building a library of movement logic functions to help with that.